Implement AuthSpider
This commit has the first working implementation of the AuthSpider. It proposes changes and assumptions:

1. The spider will ONLY log in, and scrape for URL in `start_urls`. To simplify the process even more, crawling was taken out of the equation, making it easier and separating crawling + scraping from the very beginning. In a future, if we want to crawl with login and then not logging in again, we could create a system where the session is received as input as well (cookies essentially).

2. I made some changes to the JSON configuration file proposed in #6 and #7. But that's why this is a draft and needs to be discussed.

3. Since the spider is no longer crawling, we now inherit from `Spider` instead of `CrawlSpider`.

Also, there is still work to do! 

The spider seems to be functional, however there is no executable file and I think that we will need to drastically change and simplify the spider structure, which is daunting and exciting at the same time. How? By eliminating all but the spider and login module that we've created. That is, eliminating pipelines, items, settings, etc. I think Scrapy can work without all of that, and we just need to master executing the spider from a Python script—or using `CrawlerProcess`—instead of using `scrapy crawl` or something like that.

The cost could be high, but we will simplify drastically the project (no additional enginee folders, and thus we get more flexibility and also, the simpler, the less mistakes).

That's all. Sorry for the long message, here's a potato. 

:potato: